- useLocalStorage eliminado (se viene un suplente) - useLocalStorage2 ahora es useLocalStorage (el suplente) - useLocalStorage ahora NO crea un estado, accede de forma directa al LS. - Limpieza de codigo
