+++ src/hooks/useAuth.js
@@ -1,11 +1,11 @@
 import * as Auth from "../services/auth";
-import { useLocalStorage2 } from "../hooks/useLocalStorage2";
+import { useLocalStorage } from "./";
 
 // esto podria ser un servicio de autentificacion como Firebase, AWS Cognito, o un endpoint de autentificacion de su propio backend
 
 export const useAuth = () => {
   //conexion ficticia con base de datos, accede al valor del LS con key="users"
-  const [getUsers, setUsers] = useLocalStorage2("users");
+  const [getUsers, setUsers] = useLocalStorage("users");
 
   return {
     signin: (email, password) => {
@@ -13,7 +13,7 @@ export const useAuth = () => {
       return Auth.doSignin(users, email, password);
     },
     signup: (user) => {
-      const users = getUsers();
+      const users = getUsers(); //users va a ser la informacion que exite en el LS al instante, a la cabeza
       return Auth.doSignup(users, setUsers, user);
     },
   };
